fix 96kHz audio signal playback
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@1342 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/switcher/audioswitcher/AudioSwitcher.cpp
@@ -199,7 +199,7 @@ HRESULT CAudioSwitcherFilter::Transform(IMediaSample* pIn, IMediaSample* pOut)
 	int bps = wfe->wBitsPerSample>>3;
 
 	int len = pIn->GetActualDataLength() / (bps*wfe->nChannels);
-	int lenout = len * wfeout->nSamplesPerSec / wfe->nSamplesPerSec;
+	int lenout = (UINT64)len * wfeout->nSamplesPerSec / wfe->nSamplesPerSec;
 
 	REFERENCE_TIME rtStart, rtStop;
 	if(SUCCEEDED(pIn->GetTime(&rtStart, &rtStop)))
